moved thunks folder under redux
