+++ src/redux/reducers/ui.ts
@@ -1,5 +1,5 @@
 import {createSlice, Draft} from '@reduxjs/toolkit';
-import {setRootFolder} from '@redux/reducers/thunks/setRootFolder';
+import {setRootFolder} from '@redux/thunks/setRootFolder';
 
 type UiState = {
   settingsOpened: boolean;
