+++ src/redux/thunks/setRootFolder.ts
